+++ lib/src/api/secure_random.dart
@@ -32,7 +32,7 @@ abstract class SecureRandom extends Algorithm {
   int nextUint32();
 
   /// Get a random [BigInteger] of [bitLength] bits.
-  BigInteger nextBigInteger(int bitLength);
+  BigInt nextBigInteger(int bitLength);
 
   /// Get a list of bytes of arbitrary length.
   Uint8List nextBytes(int count);
